pyproject.toml
[project]
name = "pyocr"
version = "0.1.0"
description = "Playing around with OCR"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "python-bidi>=0.6.7",
]
